インナークラスを外部化
